+++ serde_duper/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `temporal` feature to enable Temporal support for `Duration`.
+
+### Changed
+
+- Bump `duper` to 0.5.0.
+
+
 ## 0.4.1 (2025-11-25)
 
 ### Changed